Create User and inherit to Computer and Player
